aider: Added functionality for exporting Chart.js graph as a PNG image when clicking the "Exportar" button.
# Aider chat conversation:

USER: Exportar should download graph as an image in any compatible format it can be png or jpg idk

ASSISTANT: To enable the "Exportar" button to download the graph as an image in a compatible format (PNG or JPG), I will modify the `relatorio.html` file to include functionality for exporting the Chart.js graph. Here's the diff for the changes:

```diff
--- frontend/relatorio.html
+++ frontend/relatorio.html
@@ ... @@
-        <input type="button" value="Exportar">
+        <input type="button" value="Exportar" onclick="exportChart();">
@@ ... @@
+    function exportChart() {
+        const url = chart.toBase64Image(); // Get the image URL
+        const link = document.createElement('a');
+        link.href = url;
+        link.download = 'productivity-report.png'; // Set the download file name
+        link.click();
+    }
```

This modification adds an `onclick` event to the "Exportar" button that triggers the `exportChart` function. This function uses the `toBase64Image` method provided by Chart.js to get a base64 image URL of the chart, which is then used to create a downloadable link. The user can then download the chart as a PNG image named `productivity-report.png`.