aider: Modified behavior of "Salvar" button to display success message as an alert when clicked.
# Aider chat conversation:

USER: When the user saves 'Cadastrar VALOR DE LIMIAR' simplesmente retorne que foi cadastrado com sucesso

ASSISTANT:
^C KeyboardInterrupt

USER: When the user saves 'Cadastrar VALOR DE LIMIAR' simplesmente retorne que foi cadastrado com sucesso as an alert

ASSISTANT: To modify the behavior of the "Salvar" button in the `valor_limiar.html` file so that it simply returns a success message when clicked, I will update the JavaScript function associated with the button. Here's the diff for the necessary changes:

```diff
--- frontend/valor_limiar.html
+++ frontend/valor_limiar.html
@@ ... @@
-            <input type="button" value="Salvar" class="save">
+            <input type="button" value="Salvar" class="save" onclick="alert('Cadastrado com sucesso!');">
```

This modification adds an `onclick` event to the "Salvar" button that triggers a JavaScript alert stating "Cadastrado com sucesso!" when the button is clicked. This provides immediate feedback to the user that the action has been successfully completed.