aider: Updated behavior of "Salvar" button in `valor_limiar.html` to redirect to `principal.html` after displaying success message.
# Aider chat conversation:

USER: after says cadastrado com sucesso returns to principal.html

ASSISTANT: To modify the behavior of the "Salvar" button in the `valor_limiar.html` file so that it redirects to `principal.html` after displaying the success message, I will update the JavaScript function associated with the button. Here's the diff for the necessary changes:

```diff
--- frontend/valor_limiar.html
+++ frontend/valor_limiar.html
@@ ... @@
-            <input type="button" value="Salvar" class="save" onclick="alert('Cadastrado com sucesso!');">
+            <input type="button" value="Salvar" class="save" onclick="saveThreshold();">
@@ ... @@
+<script>
+function saveThreshold() {
+    alert('Cadastrado com sucesso!');
+    window.location.href = 'principal.html';
+}
+</script>
```

This modification updates the `onclick` event for the "Salvar" button to call a new JavaScript function `saveThreshold()`. This function first displays the alert "Cadastrado com sucesso!" and then redirects the user to `principal.html`.