aider: Redirect user based on fetched role from endpoint.
# Aider chat conversation:

USER: It is not working properly, it should goes to principal.html just if the user is 'Gerente' you can check that http://172.203.231.235:8080/tipo-de-usuario, that has a return like {
	"tipo_de_usuario": "Usuário"
}.    If the tipo_de_usuario is not 'Gerente' the user should be redirected to principal_convidados.html

ASSISTANT: Here are the changes to ensure that the redirection in `login.html` is based on the user role fetched from the endpoint `http://172.203.231.235:8080/tipo-de-usuario`. The redirection will occur to `principal.html` if the user is 'Gerente', otherwise it will redirect to `principal_convidados.html`.

```diff
--- frontend/login.html
+++ frontend/login.html
@@ ... @@
-                    if (data.tipo_de_usuario === 'Gerente') {
+                    fetch('http://172.203.231.235:8080/tipo-de-usuario')
+                    .then(response => response.json())
+                    .then(userData => {
+                        if (userData.tipo_de_usuario === 'Gerente') {
+                            window.location.href = 'principal.html';
+                        } else {
+                            window.location.href = 'principal_convidados.html';
+                        }
+                    })
+                    .catch(error => {
+                        console.error('Erro ao determinar o tipo de usuário:', error);
+                        alert('Erro ao conectar ao servidor para determinar o tipo de usuário.');
+                    });
```

This diff modifies the login form submission handling in `login.html` to include a fetch request to the endpoint that provides the user role. Based on the response, it redirects the user to the appropriate page.